LayoutTests/accessibility/custom-elements/liveregions-expected.txt

This tests that aria fallback roles work correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.accessibleElementById("custom-element") is null
PASS accessibilityController.accessibleElementById("custom-liveregion").boolAttributeValue("AXARIAAtomic") is true
PASS accessibilityController.accessibleElementById("custom-liveregion").boolAttributeValue("AXElementBusy") is true
PASS accessibilityController.accessibleElementById("custom-liveregion").stringAttributeValue("AXARIALive") is "polite"
PASS accessibilityController.accessibleElementById("custom-liveregion").stringAttributeValue("AXARIARelevant") is "additions"
PASS accessibilityController.accessibleElementById("text-liveregion").boolAttributeValue("AXARIAAtomic") is false
PASS accessibilityController.accessibleElementById("text-liveregion").boolAttributeValue("AXElementBusy") is false
PASS accessibilityController.accessibleElementById("text-liveregion").stringAttributeValue("AXARIALive") is "assertive"
PASS accessibilityController.accessibleElementById("text-liveregion").stringAttributeValue("AXARIARelevant") is "text"
PASS successfullyParsed is true

TEST COMPLETE
